apps/auth-svc/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0
httpx>=0.24.0
loguru>=0.7.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0